COLDCHAMBER. Light controller commands fix.

+++ libraries/LightController/src/LightController.h
@@ -43,7 +43,7 @@ public:
 
     void setMode(uint8_t m);
 
-    void setTimeout(uint16_t t);
+    void setTimeout(uint32_t t);
 
     void setActivityRatio(float ratio);
